2.13.2 Repairing -> Smiting, Salvage nerf, No more Float xp!

+++ src/main/java/harmonised/pmmo/commands/CommandClear.java
@@ -42,7 +42,7 @@ public class CommandClear
 			NetworkHandler.sendToPlayer( new MessageXp( 0f, 42069, 0f, true ), (ServerPlayerEntity) player );
 			for( String tag : keySet )
 			{
-				NetworkHandler.sendToPlayer( new MessageXp( skillsTag.getFloat( tag ), Skill.getInt( tag ), 0, true ), (ServerPlayerEntity) player );
+				NetworkHandler.sendToPlayer( new MessageXp( skillsTag.getDouble( tag ), Skill.getInt( tag ), 0, true ), (ServerPlayerEntity) player );
 			}
 
 			AttributeHandler.updateDamage( player );
